Enable symfony routing, index.html to Meot bundle.
I couldn't get any of the routing & security examples in the symfony
documentation to work. It turns out that routing requires the Apache
rewrite module to be on. They had an .htaccess in the web/ directory
that configures the rewrite module, but since it's not recommended to
use .htaccess files, Ubuntu by default does not enable .htaccess. This
means that when Apache was serving files, it wasn't using symfony
routing, but just the default static files routing. Which is why none of
the examples work.

The fix is to move the directives in the .htaccess into the Apache vhost
configuration file.

This fix broke the static file routing, so the web/index.html solution
no longer worked. Instead, I've now reconfigured routing to serve the
index from the index.html.php template in
src/Meot/FormBundle/Resources/views/Default/ instead of web/.

Note that index.html.php isn't using the default twig templating engine
as it was conflicting with Angular expressions.