Fixed position nav & question panel, iconify nav.
Making the question bank fixed position should be more convenient for
adding questions as users can just scroll to the question they want, and
then scroll to the section of the form they want, then just drag and
drop the questions in. It also fits well with the fixed position top
navbar.

Note that because there's no way for css to properly deal with fixed
position in a responsive manner and because bootstrap won't let me mix
fixed width and responsive width columns, I had to style my own fixed
width sidebar with a responsive main content area.

The print button was moved to the top navbar. The top navbar now has
icons instead of text buttons.