Fix questions being duplicated on form save.
Questions were being duplicated on save because question.is_public and
question.is_master were being set to false. This is a bit
counter-intuitive but apparently for boolean fields, symfony's data
binding treats the existence of the field as true and non-existence as
false. So even though is_master is set to false, it doesn't care about
the value, just that it exists, which means is_master is set to true.

is_master controls whether the question is a template.
is_public controls whether the question is shared.

Note that is_public only matters when is_master is true.

Also changed the "New" button on the sidebar to an icon.