Various css tweaks and fixes.
Remove question button was pushing questions down with its presence,
changed positioning so that it lives in each question's top margin.

Fix print styles having long questions overflowing the edge of the
border. Unfortunately, this seems to be caused by bootstrap, so I
limited bootstrap to only load for screens and not print.

Decrease the truncation character limit. It turns out that when the
scrollbar appears, it decreases the space available for question text.
I've added tooltips of the whole question text to each question to
compensate.

Note that I gave up on using jquery tooltips, can't seem to get them to
pick up my configuration directives, so it's just using the built in
browser tooltips right now.

The Form title now has a gray background instead of the form questions
area, it looks a bit better that way.